moved App.js from components folder to containers folder in todos example
