+++ examples/todos/index.js
@@ -4,7 +4,7 @@ import { render } from 'react-dom'
 import { Provider } from 'react-redux'
 import { createStore } from 'redux'
 import todoApp from './reducers'
-import App from './components/App'
+import App from './containers/App'
 
 let store = createStore(todoApp)
